Squash commits.
