+++ lsif/src/server/routes/dumps.ts
@@ -1,17 +1,17 @@
 import * as settings from '../settings'
 import * as validation from '../middleware/validation'
 import express from 'express'
+import { DumpManager } from '../../shared/store/dumps'
+import { extractLimitOffset } from '../pagination/limit-offset'
 import { nextLink } from '../pagination/link'
 import { wrap } from 'async-middleware'
-import { extractLimitOffset } from '../pagination/limit-offset'
-import { XrepoDatabase } from '../../shared/xrepo/xrepo'
 
 /**
  * Create a router containing the LSIF dump endpoints.
  *
- * @param xrepoDatabase The cross-repo database.
+ * @param dumpManager The dumps manager instance.
  */
-export function createDumpRouter(xrepoDatabase: XrepoDatabase): express.Router {
+export function createDumpRouter(dumpManager: DumpManager): express.Router {
     const router = express.Router()
 
     interface DumpsQueryArgs {
@@ -32,7 +32,7 @@ export function createDumpRouter(xrepoDatabase: XrepoDatabase): express.Router {
                 const { query, visibleAtTip }: DumpsQueryArgs = req.query
                 const { limit, offset } = extractLimitOffset(req.query, settings.DEFAULT_DUMP_PAGE_SIZE)
 
-                const { dumps, totalCount } = await xrepoDatabase.getDumps(
+                const { dumps, totalCount } = await dumpManager.getDumps(
                     decodeURIComponent(req.params.repository),
                     query,
                     !!visibleAtTip,
@@ -53,7 +53,7 @@ export function createDumpRouter(xrepoDatabase: XrepoDatabase): express.Router {
         '/dumps/:repository/:id([0-9]+)',
         wrap(
             async (req: express.Request, res: express.Response): Promise<void> => {
-                const dump = await xrepoDatabase.getDumpById(parseInt(req.params.id, 10))
+                const dump = await dumpManager.getDumpById(parseInt(req.params.id, 10))
                 if (dump?.repository === decodeURIComponent(req.params.repository)) {
                     res.json(dump)
                     return
@@ -68,9 +68,9 @@ export function createDumpRouter(xrepoDatabase: XrepoDatabase): express.Router {
         '/dumps/:repository/:id([0-9]+)',
         wrap(
             async (req: express.Request, res: express.Response): Promise<void> => {
-                const dump = await xrepoDatabase.getDumpById(parseInt(req.params.id, 10))
+                const dump = await dumpManager.getDumpById(parseInt(req.params.id, 10))
                 if (dump?.repository === decodeURIComponent(req.params.repository)) {
-                    await xrepoDatabase.deleteDump(dump)
+                    await dumpManager.deleteDump(dump)
                     res.status(204).send()
                     return
                 }
