+++ lsif/src/server/routes/uploads.ts
@@ -1,25 +1,25 @@
-import * as xrepoModels from '../../shared/models/xrepo'
+import * as pgModels from '../../shared/models/pg'
 import * as settings from '../settings'
 import * as validation from '../middleware/validation'
 import express from 'express'
 import { nextLink } from '../pagination/link'
 import { wrap } from 'async-middleware'
 import { extractLimitOffset } from '../pagination/limit-offset'
-import { UploadsManager } from '../../shared/uploads/uploads'
+import { UploadManager } from '../../shared/store/uploads'
 
 /**
  * Create a router containing the upload endpoints.
  *
- * @param uploadsManager The uploads manager instance.
+ * @param uploadManager The uploads manager instance.
  */
-export function createUploadRouter(uploadsManager: UploadsManager): express.Router {
+export function createUploadRouter(uploadManager: UploadManager): express.Router {
     const router = express.Router()
 
     router.get(
         '/uploads/stats',
         wrap(
             async (req: express.Request, res: express.Response): Promise<void> => {
-                res.send(await uploadsManager.getCounts())
+                res.send(await uploadManager.getCounts())
             }
         )
     )
@@ -39,8 +39,8 @@ export function createUploadRouter(uploadsManager: UploadsManager): express.Rout
             async (req: express.Request, res: express.Response): Promise<void> => {
                 const { query }: UploadsQueryArgs = req.query
                 const { limit, offset } = extractLimitOffset(req.query, settings.DEFAULT_UPLOAD_PAGE_SIZE)
-                const { uploads, totalCount } = await uploadsManager.getUploads(
-                    req.params.state as xrepoModels.LsifUploadState,
+                const { uploads, totalCount } = await uploadManager.getUploads(
+                    req.params.state as pgModels.LsifUploadState,
                     query,
                     limit,
                     offset
@@ -59,7 +59,7 @@ export function createUploadRouter(uploadsManager: UploadsManager): express.Rout
         '/uploads/:id([0-9]+)',
         wrap(
             async (req: express.Request, res: express.Response): Promise<void> => {
-                const upload = await uploadsManager.getUpload(parseInt(req.params.id, 10))
+                const upload = await uploadManager.getUpload(parseInt(req.params.id, 10))
                 if (upload) {
                     res.send(upload)
                     return
@@ -76,7 +76,7 @@ export function createUploadRouter(uploadsManager: UploadsManager): express.Rout
         '/uploads/:id([0-9]+)',
         wrap(
             async (req: express.Request, res: express.Response): Promise<void> => {
-                if (await uploadsManager.deleteUpload(parseInt(req.params.id, 10))) {
+                if (await uploadManager.deleteUpload(parseInt(req.params.id, 10))) {
                     res.status(204).send()
                     return
                 }
