+++ lsif/src/server/tasks/runner.ts
@@ -4,17 +4,17 @@ import { updateQueueSizeGauge, resetStalledUploads, cleanOldUploads, cleanFailed
 import { Connection } from 'typeorm'
 import { logAndTraceCall, TracingContext } from '../../shared/tracing'
 import { Logger } from 'winston'
-import { tryWithLock } from '../../shared/locks/locks'
-import { UploadsManager } from '../../shared/uploads/uploads'
+import { tryWithLock } from '../../shared/store/locks'
+import { UploadManager } from '../../shared/store/uploads'
 
 /**
  * Begin running cleanup tasks on a schedule in the background.
  *
  * @param connection The Postgres connection.
- * @param uploadsManager The uploads manager instance.
+ * @param uploadManager The uploads manager instance.
  * @param logger The logger instance.
  */
-export function startTasks(connection: Connection, uploadsManager: UploadsManager, logger: Logger): void {
+export function startTasks(connection: Connection, uploadManager: UploadManager, logger: Logger): void {
     /**
      * Start invoking the given task on an interval.
      *
@@ -40,12 +40,12 @@ export function startTasks(connection: Connection, uploadsManager: UploadsManage
         tryWithLock(connection, name, () => logAndTraceCall({ logger }, name, task))
 
     runTask(
-        wrapTask('Resetting stalled uploads', ctx => resetStalledUploads(uploadsManager, ctx)),
+        wrapTask('Resetting stalled uploads', ctx => resetStalledUploads(uploadManager, ctx)),
         settings.RESET_STALLED_UPLOADS_INTERVAL
     )
 
     runTask(
-        wrapTask('Cleaning old uploads', ctx => cleanOldUploads(uploadsManager, ctx)),
+        wrapTask('Cleaning old uploads', ctx => cleanOldUploads(uploadManager, ctx)),
         settings.CLEAN_OLD_UPLOADS_INTERVAL
     )
 
@@ -54,5 +54,5 @@ export function startTasks(connection: Connection, uploadsManager: UploadsManage
         settings.CLEAN_FAILED_UPLOADS_INTERVAL
     )
 
-    runTask((): Promise<void> => updateQueueSizeGauge(uploadsManager), settings.UPDATE_QUEUE_SIZE_GAUGE_INTERVAL)
+    runTask((): Promise<void> => updateQueueSizeGauge(uploadManager), settings.UPDATE_QUEUE_SIZE_GAUGE_INTERVAL)
 }
