+++ lsif/src/server/tasks/uploads.ts
@@ -5,28 +5,28 @@ import * as path from 'path'
 import * as settings from '../settings'
 import { createSilentLogger } from '../../shared/logging'
 import { TracingContext } from '../../shared/tracing'
-import { UploadsManager } from '../../shared/uploads/uploads'
+import { UploadManager } from '../../shared/store/uploads'
 
 /**
  * Update the value of the unconverted uploads gauge.
  *
- * @param uploadsManager The uploads manager instance.
+ * @param uploadManager The uploads manager instance.
  */
-export const updateQueueSizeGauge = async (uploadsManager: UploadsManager): Promise<void> =>
-    metrics.unconvertedUploadSizeGauge.set(await uploadsManager.getCount('queued'))
+export const updateQueueSizeGauge = async (uploadManager: UploadManager): Promise<void> =>
+    metrics.unconvertedUploadSizeGauge.set(await uploadManager.getCount('queued'))
 
 /**
  * Move all unlocked uploads that have been in `processing` state for longer than
  * `STALLED_UPLOAD_MAX_AGE` back to the `queued` state.
  *
- * @param uploadsManager The uploads manager instance.
+ * @param uploadManager The uploads manager instance.
  * @param ctx The tracing context.
  */
 export const resetStalledUploads = async (
-    uploadsManager: UploadsManager,
+    uploadManager: UploadManager,
     { logger = createSilentLogger() }: TracingContext
 ): Promise<void> => {
-    for (const id of await uploadsManager.resetStalled(settings.STALLED_UPLOAD_MAX_AGE)) {
+    for (const id of await uploadManager.resetStalled(settings.STALLED_UPLOAD_MAX_AGE)) {
         logger.debug('Reset stalled upload conversion', { id })
     }
 }
@@ -34,14 +34,14 @@ export const resetStalledUploads = async (
 /**
  * Remove all upload data older than `UPLOAD_MAX_AGE`.
  *
- * @param uploadsManager The uploads manager instance.
+ * @param uploadManager The uploads manager instance.
  * @param ctx The tracing context.
  */
 export const cleanOldUploads = async (
-    uploadsManager: UploadsManager,
+    uploadManager: UploadManager,
     { logger = createSilentLogger() }: TracingContext
 ): Promise<void> => {
-    const count = await uploadsManager.clean(settings.UPLOAD_MAX_AGE)
+    const count = await uploadManager.clean(settings.UPLOAD_MAX_AGE)
     if (count > 0) {
         logger.debug('Cleaned old uploads', { count })
     }
