+++ lsif/src/shared/database/postgres.ts
@@ -1,5 +1,5 @@
 import * as metrics from './metrics'
-import * as xrepoModels from '../models/xrepo'
+import * as pgModels from '../models/pg'
 import pRetry from 'p-retry'
 import { Configuration } from '../config/config'
 import { Connection, createConnection as _createConnection, EntityManager } from 'typeorm'
@@ -29,18 +29,18 @@ const MAX_SCHEMA_POLL_RETRIES = readEnvInt('MAX_SCHEMA_POLL_RETRIES', 60)
 const SCHEMA_POLL_INTERVAL = readEnvInt('SCHEMA_POLL_INTERVAL', 5)
 
 /**
- * How many times to try to connect to the cross-repository database on startup.
+ * How many times to try to connect to Postgres on startup.
  */
 const MAX_CONNECTION_RETRIES = readEnvInt('MAX_CONNECTION_RETRIES', 60)
 
 /**
- * How long to wait (in seconds) between cross-repository connection attempts.
+ * How long to wait (in seconds) between Postgres connection attempts.
  */
 const CONNECTION_RETRY_INTERVAL = readEnvInt('CONNECTION_RETRY_INTERVAL', 5)
 
 /**
  * Create a Postgres connection. This creates a typorm connection pool with the
- * name `xrepo`. The connection configuration is constructed by the method
+ * name `lsif`. The connection configuration is constructed by the method
  * `createPostgresConnectionOptions`. This method blocks (failing after a configured
  * time) until the connection is established, then blocks indefinitely while the
  * database migration state is behind the expected minimum, or dirty.
@@ -76,11 +76,11 @@ export async function createPostgresConnection(configuration: Configuration, log
 }
 
 /**
- * Create a connection to the cross-repository database. This will re-attempt to
- * access the database while the database does not exist. This is to give some
- * time to the frontend to run the migrations that create the LSIF tables. The
- * retry interval and attempt count can be tuned via `MAX_CONNECTION_RETRIES` and
- * `CONNECTION_RETRY_INTERVAL` environment variables.
+ * Create a connection to Postgres. This will re-attempt to access the database while
+ * the database does not exist. This is to give some time to the frontend to run the
+ * migrations that create the LSIF tables. The retry interval and attempt count can
+ * be tuned via `MAX_CONNECTION_RETRIES` and `CONNECTION_RETRY_INTERVAL` environment
+ * variables.
  *
  * @param connectionOptions The connection options.
  * @param logger The logger instance.
@@ -88,7 +88,7 @@ export async function createPostgresConnection(configuration: Configuration, log
 function connect(connectionOptions: PostgresConnectionCredentialsOptions, logger: Logger): Promise<Connection> {
     return pRetry(
         () => {
-            logger.debug('Connecting to cross-repository database')
+            logger.debug('Connecting to Postgres')
             return connectPostgres(connectionOptions, '')
         },
         {
@@ -101,7 +101,7 @@ function connect(connectionOptions: PostgresConnectionCredentialsOptions, logger
 }
 
 /**
- * Create a connection to the cross-repository database.
+ * Create a connection to Postgres.
  *
  * @param connectionOptions The connection options.
  * @param suffix The database suffix (used for testing).
@@ -112,8 +112,8 @@ export function connectPostgres(
 ): Promise<Connection> {
     return _createConnection({
         type: 'postgres',
-        name: `xrepo${suffix}`,
-        entities: xrepoModels.entities,
+        name: `lsif${suffix}`,
+        entities: pgModels.entities,
         logging: ['warn', 'error'],
         maxQueryExecutionTime: 1000,
         ...connectionOptions,
@@ -133,7 +133,7 @@ function waitForMigrations(connection: Connection, logger: Logger): Promise<void
 
         const version = parseInt(await getMigrationVersion(connection), 10)
         if (isNaN(version) || version < MINIMUM_MIGRATION_VERSION) {
-            throw new Error('Cross-repository database not up to date')
+            throw new Error('Postgres database not up to date')
         }
     }
 
@@ -171,7 +171,7 @@ async function getMigrationVersion(connection: Connection): Promise<string> {
  * @param callback The function invoke with the connection.
  */
 export function instrumentQuery<T>(callback: () => Promise<T>): Promise<T> {
-    return instrument(metrics.xrepoQueryDurationHistogram, metrics.xrepoQueryErrorsCounter, callback)
+    return instrument(metrics.postgresQueryDurationHistogram, metrics.postgresQueryErrorsCounter, callback)
 }
 
 /**
