+++ lsif/src/shared/store/locks.ts
